Add `ModelDataset` and move stuff around
Renamed the original `Dataset` class to `PrepDataset`.
This is because another `Dataset` class is needed that extends
torch.dataset to apply the t5 tokenizer.
The ugly part is that this part needs the indexed POS tags; so I added a
fucntion `index_data` that adds another field `self.FIRST_POS_INDEXED`
on the class (second POS and labels too). This function takes the
indexer as input to allow synchronizing between train/test/dev stages.

Also athet some internal state to the `Dataset` class to keep track
where in the pipline we are("initial", "extracted", "indexed"). After
indexed the object can be used to instantiate a `ModelDataset`.

Thank you for attending my TED Talk.